apps/camera-service/requirements.txt
# Camera Service Requirements
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-gphoto2==2.5.0
Pillow==10.2.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-multipart==0.0.6
